requirements.txt
RPi.GPIO>=0.7.0
spidev>=3.5
gpiozero>=1.6.0
Pillow>=8.0.0
numpy>=1.20.0